refactor(xtask): compile the CLI definition instead of linking huub-cli
Depending on `huub-cli` pulled in `huub`, and with it a full CaDiCaL build, just
to read Clap's command metadata. `xtask` now compiles `cli.rs` through `#[path]`
instead, which requires moving the `Lowerer` defaults, the tracing helpers, and
the `ReduceType` conversion out of that module.
